Fix #602. Fixing the imprinter and infuser GUIs to work like the alyzers.
Imprinter and infuser would previously not run their output until the output
slot was clicked (likely, only in SSP). Treealyzer and Beealyzer comment out a
block of code (Flutterlyzer never had this in the first place) that was
uncommented in these... doing the same here made the issue go away.

Sign CLA